basic chunk setup and meshing (very naive chunk meshing)
block registry implementation as well

+++ src/simulation_world/chunk/types/block.rs
@@ -1,4 +1,4 @@
-#[derive(Clone)]
+#[derive(Clone, Copy, PartialEq, Eq)]
 pub struct Block {
-    pub id: u32,
+    pub id: u16,
 }
